English_teacher/requirements.txt
# Основные библиотеки для Telegram бота
aiogram==3.4.1
python-dotenv==1.0.0

# База данных PostgreSQL
sqlalchemy==2.0.23
asyncpg==0.29.0

# HTTP клиент для OpenAI API
aiohttp==3.9.1
openai==1.35.0
httpx==0.27.0

# Планировщик задач (для автоматических уроков)
apscheduler==3.10.4

# Работа с аудио (для Text-to-Speech)
gTTS==2.4.0
pyttsx3==2.90

# Работа с файлами и временем
python-dateutil==2.8.2

# Логирование
loguru==0.7.2

# Дополнительные утилиты
pydantic==2.5.2
typing-extensions==4.9.0 